node: build_lib.js which generates node.lib
Moves/replaces the `windows_generate_node_lib.js` script
for creating the `node.lib` file needed for `zig build node_client`
to work on windows. This also modifies `build.zig` to actually create
`tb_nodeclient.dll` for windows.
